Create CONTRIBUTING.md
This folds in the instructions from README and wraps it with an attempt at the goodness needed to walk people through the contribution steps without presuming they are already familiar with GitHub and git workflow.

In the process I learned about the branchname policy and that I should have a story ID for this change. Following this commit-and-PR I will make a story for each of the bits I am working on.